--- a/nativelib/src/main/resources/scala-native/gc/shared/ScalaNativeGC.h
+++ b/nativelib/src/main/resources/scala-native/gc/shared/ScalaNativeGC.h
@@ -12,13 +12,13 @@
 #pragma comment(lib, "kernel32.lib")
 #include <windows.h>
 typedef DWORD ThreadRoutineReturnType;
-typedef ThreadRoutineReturnType(WINAPI *ThreadStartRoutine)(void *);
 #else
 #include <pthread.h>
+typedef WINAPI;
 typedef void *ThreadRoutineReturnType;
-typedef ThreadRoutineReturnType (*ThreadStartRoutine)(void *);
 #endif
 
+typedef ThreadRoutineReturnType(WINAPI *ThreadStartRoutine)(void *);
 typedef void *RoutineArgs;
 
 void scalanative_GC_init();